refactor: split constructor tests
