+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/utils/either/eitherTestHelpers.kt
@@ -3,54 +3,40 @@ package xyz.lbres.kotlinutils.utils.either
 import xyz.lbres.kotlinutils.utils.simpleIf
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
-import kotlin.test.assertIs
-import kotlin.test.assertIsNot
 import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
-/**
- * Check that all fields in [either] match the expected state when the left value is set to [value]
- */
+// Check that all fields match the expected state when the left value is set to a specific value
 fun <S, T> checkLeft(either: Either<S, T>, value: S, mutable: Boolean? = null) {
     assertEquals(value, either.left)
     assertNull(either.right)
     assertTrue(either.isLeft)
     assertFalse(either.isRight)
 
-    if (mutable == true) {
-        assertIs<MutableEither<S, T>>(either)
-    } else if (mutable == false) {
-        assertIsNot<MutableEither<S, T>>(either)
+    if (mutable != null) {
+        assertEquals(mutable, either is MutableEither<S, T>)
     }
 }
 
-/**
- * Check that all fields in [either] match the expected state when the right value is set to [value]
- */
+// Check that all fields match the expected state when the right value is set to a specific value
 fun <S, T> checkRight(either: Either<S, T>, value: T, mutable: Boolean? = null) {
     assertNull(either.left)
     assertEquals(value, either.right)
     assertFalse(either.isLeft)
     assertTrue(either.isRight)
 
-    if (mutable == true) {
-        assertIs<MutableEither<S, T>>(either)
-    } else if (mutable == false) {
-        assertIsNot<MutableEither<S, T>>(either)
+    if (mutable != null) {
+        assertEquals(mutable, either is MutableEither<S, T>)
     }
 }
 
-/**
- * Check both directions of equality for [first] and [second]
- */
+// check both directions of equality
 fun <S, T, U, V> bothEquals(first: Either<S, T>, second: Either<U, V>) {
     assertTrue(first == second)
     assertTrue(second == first)
 }
 
-/**
- * Check both directions of non-equality for [first] and [second]
- */
+// check both directions of non-equality
 fun <S, T, U, V> bothNotEquals(first: Either<S, T>, second: Either<U, V>) {
     assertFalse(first == second)
     assertFalse(second == first)
@@ -73,7 +59,7 @@ fun <S, T> setLeft(either: Either<S, T>, value: S, mutable: Boolean): Either<S, 
         either.left = value
         either
     } else {
-        Either(value)
+        Either.withLeft(value)
     }
 }
